Split worker logic into focused modules
